// src/modules/tmdb/types/tmdb.ts
export enum ContentType {
  MOVIE = 'movie',
  TV = 'tv',
}

export enum TrendingType {
  ALL = 'ALL',
  POPULAR = 'POPULAR',
  TOP_RATED = 'TOP_RATED',
  THEATHER = 'THEATHER',
}

export interface Search<T> {
  page: number;
  results: T[];
  total_pages: number;
  total_results: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface SearchMovie {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface SearchTv {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  origin_country: string[];
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  first_air_date: string;
  name: string;
  vote_average: number;
  vote_count: number;
}